firezone/custom-scripts/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
ecdsa==0.18.0
idna==3.4
pyasn1==0.4.8
python-jose==3.3.0
python-keycloak==2.6.0
python-waldur-client==0.1.5
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
